+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Ensure `@import "…" reference` never generates utilities ([#15307](https://github.com/tailwindlabs/tailwindcss/pull/15307))
+- Fix dependency related warnings when using `@tailwindcss/postcss` on Windows
 
 ## [4.0.0-beta.5] - 2024-12-04